Added moving_up and moving_down attribute to ship
Added moving_up and moving_down attribute to to Ship class.
This will be built on, and it will enable the ship to be able to move
up and down the screen. The code will be similar with that of
moving_right and moving_left in the check_events() function
under game_functions.py