Enabled the ship to move up and down
Modified the game_functions.py to include a conditional that
sets the moving_up flag to True when it detects the up button-pressed
(event) and sets the moving_down flag to True when it detects the
down key button-pressed (event). And sets the moving_up and moving_down
flag to false when it notice a release in the buttons respectively.

Modified the update method in the ship.py  file to include a conditional
that moves the ship's y value either up and down depending on the
boolean values of the moving_up and moving_down. The conditional also
ensures that the movement is limited to the screen width and screen
height.
The "self.rect.centerx = self.center
    self.rect.centery= self.ship_yaxis" ensures that it then gets
shown to the screen after the conditional is run.
Note without those lines: Though the conditional will work, it won't be
displayed.