Created a bullet Group and added arguments to fns
Created an instance called bullets from pygame.sprite.Group class
which was imported. This will be useful in treating the bullets as a
group instead of a single item.

Added new arguements to check_events and update_screen methods which
will be useful in adding bullets to the screen when the spacebar is
pressed.