Added parameters and conditional for making bullet
Added parameters to check_events() that corresponds with the arguements
in the function call already in alien_invasion.py file.

Added parameters also in the check_keydown() function that checks if
the down button is pressed, added the corresponding arguements in the
function call of check_keydown() function.

Added a conditional in the check_keydown function that detects when
a space button is pressed. The conditional contains code that creates
a bullet instance and adds it to the group of bullets whenever
tpace button is pressed.