Limited the bullets to 3 on screen at a time
created a new attribute called bullets_allowed and set it to 3 on
the settings.py file. Added a conditional in the check_key_down()
function that limits the number of bullets fired on screen to 3 at a
time.

When the bullet leaves the screen the user will be able to fire again.
But it will still be limited to 3 bullets at a time. The conditional is
set just below the conditional for the space-key-pressed.