Refactored the alien_invasion.py code
Refactored the code and moved the code lines that updates the bullet
position to a new function called update_bullets(), the update_bullets
function is now in game_functions.py, where many of the game functions
are.

The purpose of this is to keep the main game file as simple as possible.

The code was only refactored and no other change was made.