Added a conditional that allow user to easily quit
Added a conditional that allows the user to easily end the game by
pressing the "Q" key on the keyboard, instead of always having to close
the window by moving the mouse to the close icon(X symbol) at the top of
the window.

Now the user can either quit the game by moving the mouse button the
"X" icon or by simply pressing the Q letter on the keyboard.

Note: Both the uppercase and lowercase of q works fine