Modify update_screen function
Update the function by putting a "alien" parameter to its arguement
as well as adding a corresponding alien.blit_me() method that calls
the blit_me() method in the alien class.

The purpose of this change is to enable me to update the screen with
the alien image when called in the main game module.