create alien instance and display alien on screen
import Alien class to main game module, Instantiate a single alien and
finally call the update_screen() function to display alien image
on the screen.

The function call to update_screen module was modified to include an
additional arguement called "alien" so as to accurately call the
latest update_screen function(), as 'alien' is now included in the
parameter name in the update_screen function.