Treat alien as group now instead of a single item
Modified alien_invasion.py module to remove the line that instantiated
the alien. I renamed 'alien' to 'aliens' as I will now be dealing
with a group now. Changed the name from "alien" to "aliens" in the
arguement of the function call to update_screen() in the
alien_invasion.py module and changed it from "alien" to "aliens" of
parameter of the function definition of the update_screen() function
in the game_functions.py module.

Changed the method from .blit_me() to .draw(screen) in the
game_functions.py module as the draw()
method helps to display objects that are treated as groups.

Added a function call called "create_fleet()" in the alien_invasion.py
module. The function will be defined later in future commits

The purpose of these changes is to be ale to work with alien as a group
insteas of a single object. All these changes work together to achieve
this specific purpose.

Please Note: I removed the import statement of the alien ,i.e.
"from alien import Alien",
because i will not be needing to instantiate the alien in this module
(alien_invasion.py).