Created multiple aliens and determined position
defined the create_fleet() function to instantiate an alien, and
apped a for loop that makes duplicates of that alien and add it to
aliens group.

Some Maths were in the for loop that helped me know where to position
each alien correctly on the horizontal parts of the screen, so it
does not appear randomly.

Please Note: I had to import the Alien class here
i.e. from alien import Alien,
as i had to work with the instance of the Alien and then made duplicates