Duplicate Alien multiple times across several rows
Added function that determines the maximum number of rows alien can
occupy on the verctical screen (get_number_rows function).

Modified the create_fleet() function to include a for loop that
duplicates the alien image horizontally till it reaches the end of the
screen-width and then, it repeats the duplication on the next row
provided that the maximum number of rows the aliens can occupy is not
reached.

The first for loop ensures it does the duplication row-by-row.
The nested for loop ensures the image (alien) gets duplicated to the
right for each row, provided that the maximum number of rows is not
exceeded.

Please Note: I added row_number as a parameter to the create_alien
function  because I know that the for loop will need it for the
duplication to be successful.

I also added ship_1 to the create_fleet function  definition
because i know that the height of the ship will be needed.
(Note that i labelled my ship as ship_1 right from the beginning)

The call to the create_fleet function is in the main game module
which is (alien_invasion.py) and it will be updated in future commits
to include the arguement "ship_1", so as to ensure that create_fleet
function runs without errors.

ALL THESE CHANGES WILL NOW ENSURE THE ALIENS THAT APPEAR HORIZONTALLY
IS NOW DUPLICATED MULTIPLE TIMES DOWN THE SCREEN. The previous code
only show 1 row of the aliens