Updated the call to create_fleet() function
The call to create fleet() function is now Updated to include "ship_1"
as an arguement, as the height of the ship is needed to help me
determine the available vertical distance( i.e.available number of rows)
.  The maths and the calculations were done in the immediate previous
commmit.

This function call is in the main game module (alien_invasion.py)

The call to this function will now display the entire alien fleet across
several rows on the screen.