Added the same ship with a smaller height
Added the main game character (the red rocket), but I made it smaller
using an editing tool. The purpose is to create enough space on the
screen. Please note that the name of this ship image is "ship5.png.

In future commit, I will modify the ship class to take this ship image
for the ship, instead of the former image named "ship.png"