Added alien speed factor attribute to Settings()
Added speed_factor attribute of the alien to the Settings class, which
will be needed to move the alien to the right and to the left.
Added it to settings class for proper organization  so all the crucial
elements of the game can be controlled there.

Added update method() to the alien class which will automatically
update the alien's position by moving the fleet to the right by the
value of the alien_speed_factor.