Move Alien fleet to the right
Added update_aliens() function to game_function module, which
will call the update method of the alien class and move the alien
to the right by the value of the alien speed factor.

Added the function call of update_aliens() which will call the update
method of the alien class and thereby move the alien to the right.