Move Alien fleet to the right
Added update_aliens() function to game_function module, which
will call the update method of the alien class and move the alien
to the right by the value of the alien speed factor.

Added the function call of update_aliens() which will call the update
method of the alien class and thereby move the alien to the right.

Note there was an error in this commit previously. I had to
rewrite this exact content of this commit for clarity sake.
For reference to the original error, kindly send an email.

So the future fix of "self.x =float(self.rect.x)"
has already been fixed here. Only this exact commit was rewritten
in the commit history. No major difference, but just addition of
"self.x =float(self.rect.x)"