Change direction of alien fleet
Added an attribute fleet direction for the aliens, to Settings().
This attribute will be able to change the direction of the alien.
You  will find this change in the settings.py module.

The speed_factor of the alien will get multiplied by the positive
value of the fleet direction (1) if alien needs to moved right, and
it will be multiplied by the negative (-1) if aliens needs to be moved
left. You will find this change in the alien.py module, inside the
update method.