Move alien fleet down when it reaches the edge
Added update_aliens() function to the game_functions() module which
will move the alien down by certain units(specified in the dropdown
attribute) and then move the alien sideways to the right or left
depending on whether the fleet hits right edge or left edge of the
screen.

Added other functions that assists the update_aliens() function.
Update alien calls the functions indirectly to perform this operation.
The update alien function calls check_fleet_edges() to
check whether the alien fleet has hit any of the screen edges.
(The check is continuous as it is in the main game while loop, check
alien_invasion.py for details).

Back to the game_functions module, when check_edges returns True, i.e.
alien fleet has hit the edge, change_fleet_direction() function is
called, which will eventually move the ship to the opposite
direction at which it came.

Finally update_aliens function is called in alien_invasion.py , after
the call to update_bullet function in the main while loop of the game.