Kill alien after bullet Hit
Added a line of code for collision detection, it checks the two distinct
groups (bullets and aliens) and detects whether there is a hit.
If true, it destroys the alien and the bullet that hits the alien.

Full breakdown
-------------

Added the "pygame.sprite.groupcollide()" code to update_bullets()
function in game_functions.py. Also updated the bullet arguement to
include "aliens" to make detection possible.

Updated the function call to update_bullets in the main game loop
which is in alien_invasion.py to incude a new parameter for aliens.