Added code to track the number of ships left
Populated the GameStats class with code to check the number of
ships left. The number of ship attribute (ship_limit) is initialized
in the settings.py module, it is found under ship settings, the ship
limit is set to 3, and the GameStats class will track the number of
times the ship has collided with the alien.