Restart game after ship to alien collision
Modified the update_aliens() function to include a ship_hit()
function that performs various operations when it detects a collision
between the ship and any alien.

First operation
The ship_hit() function first
reduces the number of ship left by 1 ( for example, a player has
default ships of 3, after first collision it gets to 2 ship left)

Second operation
It emties the aliens and bullet that were on the screen at the time
of collision

Third operation
It creates a new fleet of aliens and place the ship back to the
default start position (bottom center). BUT IT DOESN'T DRAW THE NEW
ALIEN FLEET AND SHIP'S POSITION TO THE SCREEN YET.

Fourth operation
It pauses the screen for a short period of time to show the ship
colliding with the alien.  The next code in the game main loop will then
draw the new fleet after screen has been paused momentarily.

So the visible output of these code is just a pause after collision.
The game would look as if a new fleet was created after the collision.
But in reality it was actually created during collision but displayed
after.