Restart game when any alien hits screen bottom
Modified the update_alien function in game_functions.py to include
a conditional that checks to see if any alien has reached the bottom
of the screen. The conditional will include a call to ship_hit
function when it detects that an alien has reached the bottom
(this simply means that the player lost that round and was not
able to shoot down all the aliens).

When the conditonal detects that an alien has reached the bottom, it
calls the ship_hit function which does the same operations when a ship
collides with an alien.
The four operations are:
1. Decrement the number of ship left by 1
2. Empty all aliens and ships remaining on screen at collision
3. Create a new alien fleet and reset the position of ship to bottom
center
4. Pause the game for a short moment before the new change take place
(i.e the change i talked about in number 2 and 3).