End game when player fails three times
Implemented the game end criteria by setting a flag called
game_active in the game_stats class. I set it to True by default,
and the flag turns false when the ship_limit is now 0, which means
that the player does not have any ship left to use. The game then
freezes.

Step-by-Step change

-Indented some codes in the main game loop (alien_invasion.py) to
fall under the conditional for the game_active flag. This means that
the  affected part of code will only run as long as the game_active flag
is set to True. It turns false when the ship limit is now 0.
The game freezes and the player will need to restart the program to
play again when the game_active flag is False.

-The game_active flag was initialized in the GameStats class inside the
game_stats.py module.  (game_stats.py)
-The ship_hit function inside the game_functions is also indented.
The change made inside this function is what sets the game_active
flag to false when the ship_limit has been reached. (game_functions.py)
-The change made in the indented main loop is what ensures that the
game is inactive when it notices that the game_active flag is now set to
False. (alien_invasion.py)