Populated the button module
Populated the button module with a Button class that creates
a button. The class contains the initializations of button attributes
including colour and size, the class also contains methods that
fills the button colour to the centre of screen and then prepare
a message and displays it on the button.