Ensured the Play button restarts the game
Added code blocks to check_events function to check for mouse clicks and
then check specifically for clicks on button area. When the button area
is clicked . i.e. play button is clicked, the game_active flag is
automatically set to True. This means that game now starts whenever
the Play button is clicked.