Added code to ensure the game resets properly
Added codes that ensures that game statistics is reset as well as
ensuring that the alien fleet is recreated when the the player fails.

The previous code only ensures that the Play button does it work, I
noticed that the button wasn't doing a reset when the user finally fails
four times.