Let playbutton region unresponsive when gameactive
The change in this Commit Ensures that if the region for play button is
tapped when game is active, the game should not reset.  Although the
button disappears when the game is active, the game still resets when
the area for the button was pressed, you can test this in the previous
commit before this commit.

This commit ensures that the Play_button code can only run when the
Mouse down button is pressed and when the game is not active.