Fixed an off-by-one error
The game should let player play three times if ship limit is set to
three, but instead it lets the player play four times. Noticed it was
because of the position of ship_hit function.

Previously, when i test the game by letting the player loose three times
, the outcome is as follows:

The player already looses 3 times but it gives the player a fourth
chance of playing because when the ship left (ship_limit in this case)
is already at 0, the game does not detect it yet because a ship hit
does not happen at that time until the next hit, so the normal game loop
runs. The game then detects it as 0 when a next hit happen after the
three ships is exhausted. as 0 is not greater than 0,
In reality the conditional of if ship_limit >0 runs four times as
ship_hit function has to be called four times before it detects that
the ships are exhausted. giving the player a total of four trials

 What I want is to check for hits to be three. The ship_hit must only
 happen three times .

 Maths to explain this

 ships_left =3

 first hit happens (reduce ship left by 1)
 ships_left =2

Second hit happens (reduce ship left by 1)
ships_left =1

Third hit happens (reduce ship left by 1)
ships_left =0

Fourth hit
It doesn't run but it sets game active flag to false

At ships_left=0, game does not become inactive yet because the ship_hit
function has to be called at ship_hit = 0 before it turns inactive,
in other words a fourth hit needs to happen for the game to be inactive
as it is the ship_hit function that turns the game inactive when it
detects that the ships left is not greater than 0.

This error arose because of the position of the conditional. To Fix
this i need the conditional to check if the ships_left is greater than
1. So at 0, this way only 3 hits happens.

ships_left =3

first hit
ships_left =2

second hit
ships_left =1

third hit
It doesn't run but it sets game active flag to false.