Refactored check_play_button function
Refactored the function so that the main code that starts the
game and does the reset is in a new function, which is called inside
the check_play_button function when the function detects that
the button was pressed.

Doing this will also help me call this new function when a player
enters a key that has the ability to start the game.