Start game when the key "P" is pressed
I added a conditional in the check events code block that checks to see
if the key "P" is pressed. When the Key is pressed, it does the same
function as pressing the button.

Now the player has the option to start the game by either pressing
the 'Play' button or pressing P on their Keyboard