Make changes to make the game more lively
Split some settings of the game into static ones and dynamic ones.
The dynamic settings will change as the game progresses while the
static ones will stay the same.

For example; the aliens need to move faster as the level progresses.
When the player has finished shooting down a fleet of aliens, the next
fleet should move faster instead of using the same speed. This
type of change will make the game more lively and engaging.

What i did in this commit is to leave the init method of settings class
to contain initializations of static settings, then call the
initializations of dynamic settings when the init method of static
settings has been called.

Future commits will contain changes that will help me achieve this.