Level up when player shoots down entire fleet
Increase the speed of the alien fleet when the player shoots down
a fleet of alien. It levels up after each fleet was shot down.
It continues levelling up until the game is restart.

In future commit, I will ensure that the play button resets the dynamic
settings of the game so it does not just continue levelling up till
the program restarts.

What i did in this commit. I created a method called increase_speed()
in the settings class that increases the speed of the dynamic settings.
This change can be seen in settings.py module

I now ensured that this method is called whenever the fleet was shot
down(which is the section of code that checks if len(alien) == 0).
This change can be seen in game_functions.py module.