Change default point for bullet hit on new levels
Change the default point that a player gets whenever they shoot an
alien on different levels. The points should get higher per level.
For example; a player can get 50 point for shooting down a alien in
level 1, the player should get 75 point for shooting an alien in
level 2. It is not fair for the player to get 50 points for all shot to
aliens in all levels because the difficulty of the game increases as
the level increases.