Formatted the score
I ensured that the score at the right hand side is first converted to
the nearest tens,hundreds and so on.

And finally i ensured that a comma separates the numbers when it is
greater than three digits.

The game should now show the numbers separated by commas if a player
has gotten points greater than the digit 999. i.e Player shoud see
1,000