Implement scoring system -Track High Score
Modified game_stats.py to include a tracking attribute called
self.high_score to track the High score of the game.

Modified score_board to include a method called prepare_High_score
which ends up taking the tracked High score, rounds it off, converts
it to string, then converts it to image, and then position it at the
top center of the screen.

Finally i added a line to the show_score method of the score_board
that ensures that the high score is now displayed at the top center of
the screen.

The High score is left static at the  moment. In future commits, i
will ensure the High score updates itself.