Reset tracking variables on Button click or key P
Reset all tracking variables when user click the mouse cursor on
play button or presses the 'P' key. Everything resets except the
HighScore because the HighScore is initialized statically in the init
method.

Reason, we do not really need the High score to always reset because
the user always need to see the High score in other to beat their
previous records.