Ship needs to decrease
Changed the tracking variable that was used from ship_limit to
ships_left because the ships_left decreases. This implies that the ships
left is what determines the number of ship images that is displayed on
screen. If i used ship_limit, the images will always stay constant
because that variable is set to stay the same throughout the program
run.