Play the sound when bullet is shot
Created instance of the sound class in alien_invasion module
(main game loop) and ensured that sound plays only when the spacebar key
and bullet on screen is within its limit
For example; at the moment the limit is 3, which means 3 bullets on
screen at a time. The sound can play 3 times as long as three bullets
are on the screen. It only plays when a bullet is released.

I did this by putting the call to the method in the check spacebar
event and then under a function with a condition that limits bullet.