Restructured the sound class and addded new method
I ensured that self.sound was made to be an attribute in the init
method of the class, as this would help me use it repeatedly in
the other methods of the class. Created a new method that reduces
the sound volume. I added this new method because the explosion
sound file i downloaded is too loud so this method will reduce the
volume.