Background music plays when game is active
Imported the class for the Background music to the main game module
(alien_invasion.py) and instantiated it to take the background music
file.

I then ensured that the background music is played when the game is
active and stops when the game is no longer active (when the ships
have finished). I did this by calling the play_music method and
calling the stop_music method in the appropriate section of the
game_functions code.