load music whenever play method is called
The load call "pygame.mixer.music.load(self.filename)" needs to be
in the play method and not the init method because pygame.mixer.music
is a shared resources and it can only stream one music at a time.
It can never be an attribute because it is not unique to a class
as pygame.mixer.music breaks Object-Oriented programming rules.