Added a new method save_and_stop_music
This method should save the timestamp af the exact point a music stops.
The attribute that should hold this is self.music_time_stamp.

The purpose of this change is to ensure that my code supports my design.
I plan to have a start music playing as long as the play button is not
clicked. Whenever the game starts (button clicked),
i plan that a different music plays.
And whenever the game ends i plan that the music that was playing
at the start continues playing from the initial point that it stopped
before the game started.