Play the game off track from where it left off
I created an instance of Background music and passed the game off
track as the file name. I then ensured that the music starts
immediately the game begins by placing it in def run_game() function,
the part just before the game loop begins, and immediately before
the alien fleet is created.

I then ensured that the music stops and save the exit timestamp
when a player presses the playbutton or key "P".

The major change to this is this line of code
>>>calm_music.save_and_stop_music()

Finally i ensured that the music starts from where it left off
whenever the player looses.

The major change to this was
>>>calm_music.play_music()

All of these changes involved me having to modify the respective
functions like check_events that checks if player presses button and
the refactored functions under check events. It also made me
modify the update_aliens function so as to ensure that the game off
track starts from where it left off whenever the game ends. The
refactored functions under update_aliens function was also modified.

It is useful to note that the game off track is inside an instance
object called calm_music.