Added a safety check to fix an Intermittent bug
Created a  try block in the play method that ensures that the music
starts off from where it left off "accurately".There was an Intermittent
bug that occured on rare occasions when the program was run and
it happened whenever the game ends, and exactly when the game off
track needs to start off from where it left.

The problem probably occured because the timestamp picked an
abnormal point in the music at the time it was saved. For example,
the point interval before the music loops again.

Added a safety check that checks if the value of the timestamp is
accurate, then if not, it ensures the music starts from point 0
(beginning) instead of the left off time.

There was also an except block which assumed the error occured from
another unknown source which is very unlikely.

In summary, this change ensures that when the error occurs,
the game off track starts from
the beginning instead of the left off point that was initially in my
plan.

Note: Since this is a rare and Intermittent bug, it is very unlikely
to happen. The game off track should always start off instead
of having to start again (according to my design).