Make the top left ships smaller
This change ensures that the ships that appear at the top left of
the screen appears smaller than the original ship size.

This is good as it makes it look like a realistic indicator of the
number of ships remaining.

I implemented this by adding a new method to the ship class that makes
the ship smaller and then calling the method of the instance that is
relevant to updating ships left statistics.

I also ensured that the space between the ships at the top left is
updated thereby ensuring that the ships appear closer to each other.