Reduced the file size for background music
This change is also for web compatibility. Though this does not
cause error, but all large files in the game needs to be optimized so
that the load time is faster. At the moment the game has an initial load
time of >5 minutes. Thats the reason for this change -to reduce the
load time.

The same file will be added again with the same name, so the files will
appear normal in this change but in reality it has been compressed,
the change will not be very noticeable when it comes to gameplay
experience as the player will still hear quality music.