Increase the speed of all elements for web
when tested, the game appears very slow on the mobile browser.
Changed the values to a nice incremented amount so game appears
much faster than the rendered type.

Please Note: These values are only for web version and not for local
repository testing.  The web version was much slower, thats the reason
why i incremented the values. If you want to test the released
repository folder, Please download it from the master branch under
the releases tab to get the released version which is
"Version 1.0.0 - Complete Release".