Rebuild the game so apk file updates speed change
Successfully build the game using pygbag so that the apk
produced contains the most updated file of settings.
Now the bullet appears to be faster when run on web.

Please note that the speed and other behaviours will
vary if game is tested locally or on the mobile browser.
This is because pygbag uses a lot of external tools to
ensure game is run successfully which can result in
game having to slow down manually. I made this change
to adapt to  the web behaviour.

Reality: web version will appear fast, local version of
the branch will appear very fast.